Conflicts:
	GenDoc/MergeDoc/Program.cs

+++ DocContextMenuPackage/DocContextMenuPackage/MDGenerateHelper.cs
@@ -14,7 +14,7 @@ namespace Company.DocContextMenuPackage
 
         public MDGenerateHelper(DTE dte, Project project, ProjectItem docsFolder, ProjectItem projectDocFolder)
         {
-             _dte = dte;
+            _dte = dte;
             _project = project;
             _docsFolder = docsFolder;
             _projectDocFolder = projectDocFolder;
@@ -28,25 +28,34 @@ namespace Company.DocContextMenuPackage
             string docMetadataPath = Path.GetDirectoryName(_docsFolder.Properties.Item("FullPath").Value.ToString());
             string arguments = string.Format("\"{0}\" /o:\"{1}\" /p:\"{2}\" /t:\"Markdown\"", _dte.Solution.FullName, docMetadataPath, projectName);
 
-            await ProcessUtility.ExecuteWin32ProcessAsync(executorPath, arguments, workingDirectory, 100000);
-            foreach (string file in Directory.GetFiles(docMetadataPath + "\\" + _project.Name))
+            var processingDetail = await ProcessUtility.ExecuteWin32ProcessAsync(executorPath, arguments, workingDirectory, 100000);
+
+            if (processingDetail.ExitCode != 0)
             {
-               File.Delete(file);
+                throw new System.ApplicationException(string.Format("Error executing {0} {1} : {2}", executorPath, arguments, processingDetail.StandardOutput + processingDetail.StandardError));
+            }
+
+            string projectDocfolderPath = Path.GetDirectoryName(_projectDocFolder.Properties.Item("FullPath").Value.ToString());
+            foreach (string file in Directory.GetFiles(projectDocfolderPath))
+            {
+                File.Delete(file);
             }
 
             int itemCount = _projectDocFolder.ProjectItems.Count;
-            for (int i=itemCount; i>=1; i--)
+            for (int i = itemCount; i >= 1; i--)
             {
-               _projectDocFolder.ProjectItems.Item(i).Remove();
+                _projectDocFolder.ProjectItems.Item(i).Remove();
             }
-            foreach (string dir in Directory.GetDirectories(docMetadataPath + "\\mdtoc\\" + _project.Name))
+
+            string tempMDPath = docMetadataPath + "\\mdtoc\\" + _project.Properties.Item("AssemblyName").Value.ToString();
+            foreach (string dir in Directory.GetDirectories(tempMDPath))
             {
                 foreach (string file in Directory.GetFiles(dir))
                 {
                     _projectDocFolder.ProjectItems.AddFromFile(file);
                 }
             }
-      }
+        }
 
-   }
+    }
 }
